broke the contact.h and contact.cpp file into seperate files for each contact type as this file was becoming massive, and difficult to read.
