+++ include/contact/contact.h
@@ -27,6 +27,7 @@
   \brief Defines the contact class hierarchy
  */
 #ifndef CONTACT_HXX
+#define CONTACT_HXX
 
 #include <Inventor/nodes/SoSeparator.h>
 #include <list>
@@ -39,6 +40,7 @@ class Body;
 class SoSeparator;
 class SoMaterial;
 class Matrix;
+class VirtualContact;
 
 #ifdef ARIZONA_PROJECT_ENABLED
 #include <arizona/Arizona_Raw_Exp.h>
@@ -94,6 +96,7 @@ struct Wrench {
  */
 class Contact
 {
+
 friend class VirtualContact;
 
 protected:
@@ -107,7 +110,7 @@ protected:
   Contact *mate;
 
   //! Coefficient of static friction
-  double cof;
+  double sCof;
 
   //! Coefficient of kinetic friction
   double kcof;
@@ -191,7 +194,7 @@ public:
   int numFCWrenches;
 
   //! Initializes an empty contact (not really used)
-  Contact() : body1(NULL),body2(NULL),mate(NULL),cof(0.0),
+  Contact() : body1(NULL),body2(NULL),mate(NULL),sCof(0.0),
 	contactForcePointers(NULL), optimalCoeffs(NULL), prevBetas(NULL), wrench(NULL) {}
 
   //! Constructs a contact between two bodies 
@@ -343,234 +346,6 @@ public:
   float coneR, coneG, coneB;
 };
 
-//! A Point Contact With Friction (PCWF) implementing a Coulomb friction model
-/*! The Point Contact simulates rigid bodies in contact. As such, its wrench
-	space is a 3D cone and its friction space is a 2D circle with just 
-	tangential friction. Its visual indicator is the cone itself. The cone's
-	angle is equal to the tangent of the friction coefficient.
-*/
-class PointContact : public Contact
-{
-public:
-	//! Also sets up friction edges according to Coulomb model
-	PointContact(Body *b1, Body *b2, position pos, vec3 norm);
-	//! Stub destructor
-	~PointContact();
-	//! Defines a 2D friction circle with tangential friction only.
-	int setUpFrictionEdges(bool dynamicsOn = false);
-	//! Returns the visual indicator which is the wrench space cone itself
-	SoSeparator* getVisualIndicator();
-};
-
-//! Soft Contact implements an SFC model for contacts between soft bodies
-/*! The SoftContact attempts to capture some of the frictional effects between
-	soft bodies without explicitly computing the deformation at the contacts.
-	It locally fits analytical surfaces to both bodies involved by computing
-	their local radii of curvature. After this, it uses analytical models
-	to compute the expected contact area and pressure distribution.
-	
-	Based on the contact area, it then tries to approximate the relationship
-	between tangential friction and frictional torque using the limit surface,
-	which becomes a 3D friction ellipsoid. The friction space is thus 3D, and
-	the wrench space becomes 4D.
-	
-	Its visual indicator shows the analytical surface that has been fit to both
-	bodies involved, as a small patch around the contact location.
-*/
-class SoftContact : public Contact
-{
-protected:
-	//! A list of points from body1 that surround the contact in the frame of body1
-	vec3 *bodyNghbd;
-
-	//! The number of points in the current fit for analytical surfaces
-	int numPts;
-
-	//! Second order fit of points from body1 in the form r1x^2 + r2y^2
-	double r1, r2;
-	//! Parameters of unrotated fit in form z= ax^2+by^2+cxy
-	double a, b, c;
-
-	//! The relative angle between r1 and the mate's r1
-	double relPhi;
-
-	//! The rotation from the contact frame to the frame in which the fit is given by the radii of curvature, r1 and r2
-	mat3 fitRot;
-
-	//! The angle of the fitRot rotation
-	double fitRotAngle;
-
-	//! The major axis of the ellipse of contact
-	double majorAxis;
-	//! The minor axis of the ellipse of contact
-	double minorAxis;
-
-	//! The relative radii of curvature of the two bodies
-	double r1prime;
-	//! The relative radii of curvature of the two bodies
-	double r2prime;
-
-	//calculates the relative angle between the frames with 2 radii of curvature between
-	//the contact and its mate
-	int CalcRelPhi();
-
-	//! Calculate the relative curvatures of the two bodies for contact dynamics
-	int CalcRprimes();
-
-	//! Fits an analytical surface to a local patch on body1 around the contact
-	void FitPoints( );
-
-	//! Calculates friction characteristics using a Mattress model
-	double CalcContact_Mattress( double nForce );
-
-public:
-	//! Also takes a local neighborhood of points around body 1
-	SoftContact( Body *b1, Body *b2, position pos, vec3 norm, Neighborhood *bn );
-
-	//! Deletes its local record of the body neighborhood
-	~SoftContact( );
-
-	//! Friction model is a 3D friction ellipsoid also containing frictional torque
-	int setUpFrictionEdges(bool dynamicsOn = false );
-
-	//! Visual indicator is a small patch of the fit analytical surface on the body
-	SoSeparator* getVisualIndicator();
-
-	//! Also attempt to apply some torques in the contact plane; currently disabled.
-	virtual void computeWrenches();
-};
-
-//! A contact that exists even when a hand is not perfectly touching another object
-/*!	This class is meant for studing grasp quality when there is really no grasp, as
-	the hand is not exactly touching an object. The high-level purpose is to convert
-	the grasp quality function to a continous function that exists in all hand
-	configurations, rather then a highly discrete function that only is non-zero
-	when the hand is touching an object in the right way. However, this line of
-	work is not completed and this class is in bad need of an overhaul.
-	
-	A virtual contact does NOT come in a pair - there is no mate. As such, any instance
-	of this will have the mate set to NULL, a distinct sign of a virtual contact. 
-	Also, virtual contacts have the normals pointing outward, instead of the usual
-	contacts whose normals point inward. This is confusing and should probably be
-	changed at some point.
-	
-	The VirtualContact does not always need a body2. In fact, most of its design can
-	work if the body2 is NULL. However, it can have a body2, in which case some
-	distance functions are computed based on the body2. This is not very well
-	implemented right now and subject to change.
-	
-	A virtual contact can be instantiated from any kind of contact as it pretty much 
-	only copies the friction edges and location. It also needs a separate 
-	computeWrenches to eliminate the need for the object.
-	
-	All computations for a virtual contact (like wrenches, etc) are performed in the
-	world coordinate system, unlike the usual contact which is in the object's 
-	coordinate system. The reason is that we can have a grasp defined my many virtual
-	contacts on the hand (rather then meny traditional contacts on an object). In this 
-	case, each link of the hand has its own coordinate frame, so in order to have a 
-	coherent frame for the grasp we use world coordinates.
-*/
-class VirtualContact : public Contact 
-{
-private:
-	//! Meant to replace the object CoG when a virtual contact has no object
-	position mCenter;
-	//! Meant to replace the object max radius when a virtual contact has no object
-	float mMaxRadius;
-	
-	//! We can hook this up directly to the world root to see the contact; for debug purposes
-	SoSeparator *mWorldInd;
-	//! We keep a pointer so we can change this contact's appearance on the fly; for debug
-	SoMaterial *mZaxisMat;
-
-	//! Which finger (kinematic chain) of the robot this virtual contact is on. -1 means the palm.
-	int mFingerNum;
-	//! Which link of the chain this virtual contact is on.
-	int mLinkNum;
-
-	//! Empty contact initialization shared by all constructors; sets the contact as its own mate
-	void init();
-public:
-	
-	//! The constructor initializes an empty virtual contact, setting it as its own mate
-	VirtualContact();
-	//! Also flips the normal of the original contact, to make it point outward
-	VirtualContact(int f, int l, Contact *original);
-	//! Copy constructor copies friction edges and location
-	VirtualContact(const VirtualContact *original);
-	//! Destructor
-	~VirtualContact();
-
-	//! Sets the body that this contact is on, presumably to a robot link
-	void setBody(Body *b){body1 = b;}
-	//! Sets an object that this contact is not exactly on, but we use in calculations.
-	void setObject(Body *b){body2 = b;}
-
-	//! Writes this contact, including friction edges, to a file
-    // Returns false if file cannot be opened for writing
-	void writeToFile(std::ofstream& outfile);
-	//! Loads this contact, including friction edges, from a file
-    // Returns false if file cannot be opened for reading 
-	bool readFromFile(std::ifstream& infile);
-
-	//! Wrench computation is done in world coordinates considers the fact that we have no object
-	void computeWrenches(bool useObjectData = false, bool simply = false);
-	
-	//! Scales the wrench space of this contact by a given factor
-	void scaleWrenches(double factor);
-
-	//! The virtual contact can not set up its own friction edges.
-	int setUpFrictionEdges(bool dynamicsOn = false);
-
-	//! Updates mObjectDistance and mObjectNormal based on current world locations.
-	void updateContact(Body* object);
-
-	//! Sets the max radius that is to be used in wrench computations
-	void setRadius(float r){mMaxRadius = r;}
-	//! Sets the c.o.g that is to be used in wrench computations
-	void setCenter(position c){mCenter = c;}
-	//! Gets the max radius that is used in wrench computations
-	double getMaxRadius(){return mMaxRadius;}
-	//! Gets the center that is used in wrench computations
-	position getCenter(){return mCenter;}
-
-	//! The visual indicator is just a thin red cylinder
-	SoSeparator* getVisualIndicator();
-	//! Another indicator that can be atatched directly to the world root, for debug
-	void getWorldIndicator(bool useObjectData = false);
-
-	//! The location of this contact in world coordinates
-	position getWorldLocation();
-	//! The normal of this contact (pointing outwards) in world coordinates
-	vec3 getWorldNormal();
-	//! Computes the distance to the object along the contact normal and the object surface normal at the closest point
-	void getObjectDistanceAndNormal(Body *body, vec3 *objDistance, vec3 *objNormal);
-
-	//! Changes the visual marker to show that this contact is marked; for debug purposes.
-	void mark(bool m);
-
-	//! Returns the number of the finger that this contact is on
-	int getFingerNum(){return mFingerNum;}
-	//! Returns the number of the link that this contact is on
-	int getLinkNum(){return mLinkNum;}
-
-        //! Changes the frame (and thus also the location and normal) of this virtual contact
-        void changeFrame(transf tr);
-};
-/* just like what VirtualContact class has done, VirtualContactOnObject Class is nothing weird except that
-it changes the virtual contacts' loaction from the finger to an object imported before.
-with these virtual contacts, we continue to study the grasp quality. Hao 10.04
-*/
-class VirtualContactOnObject : public VirtualContact
-{
-public:
-	VirtualContactOnObject ();
-	~VirtualContactOnObject ();
-	bool readFromFile(std::ifstream& infile);
-#ifdef ARIZONA_PROJECT_ENABLED
-	void readFromRawData(ArizonaRawExp* are, QString file, int index, bool flipNormal = false);
 #endif
-	void writeToFile(std::ofstream& outfile);
-};
-#define CONTACT_HXX
-#endif
+
+
