+++ src/body.cpp
@@ -47,6 +47,8 @@
 #include "dynJoint.h"
 #include "ivmgr.h"
 #include "contact.h"
+#include "virtualContact.h"
+#include "virtualContactOnObject.h"
 #include "graspitGUI.h"
 #include "tinyxml.h"
