+++ src/contact/virtualContactOnObject.cpp
@@ -1,0 +1,185 @@
+#include "virtualContactOnObject.h"
+
+#include "contact/contact.h"
+#include "debug.h"
+#include "body.h"
+#include "world.h"
+
+#include <Inventor/nodes/SoCone.h>
+#include <Inventor/nodes/SoCoordinate3.h>
+#include <Inventor/nodes/SoCylinder.h>
+#include <Inventor/nodes/SoIndexedFaceSet.h>
+#include <Inventor/nodes/SoMaterial.h>
+#include <Inventor/nodes/SoTransform.h>
+#include <Inventor/nodes/SoTranslation.h>
+#include <Inventor/nodes/SoSeparator.h>
+#include <Inventor/nodes/SoSphere.h>
+
+#include <iostream>
+#include <fstream>
+
+VirtualContactOnObject::VirtualContactOnObject()
+{
+    wrench = NULL;
+    body2 = NULL;
+    mate = this;
+    prevBetas = NULL;
+}
+
+VirtualContactOnObject::~VirtualContactOnObject()
+{
+}
+
+bool
+VirtualContactOnObject::readFromFile(std::ifstream& inFile)
+{
+    if (!inFile.is_open())
+    {
+      DBGA("VirtualContact::readFromFile - Failed to read from file");
+      return false;
+    }
+
+    float w,x,y,z;
+
+    //numFCVectors
+    inFile >> numFrictionEdges;
+    if (inFile.fail()){
+      DBGA("VirtualContactOnObject::readFromFile - Failed to read number of friction vectors");
+      return false;
+    }
+
+    //frictionEdges
+    for (int i=0; i<numFrictionEdges; i++) {
+        for (int j=0; j<6; j++) {
+          inFile >> w;
+          if(inFile.fail()){
+            DBGA("VirtualContactOnObject::readFromFile - Failed to read number of friction edges");
+            return false;
+          }
+          frictionEdges[6*i+j] = w;
+        }
+    }
+    fprintf(stderr,"\n<frictionEdges scanned successfully>"); // for test
+
+    // (w,x,y,z) is already a quaternion, if you want to do frame rotate v rad along a vector (x,y,z),
+    //you can use q(v,vec(x,y,z))
+    Quaternion q;
+    vec3 t;
+    inFile >> w >> x >> y >> z;
+    if(inFile.fail()) {
+      DBGA("VirtualContactOnObject::readFromFile - Failed to read virtual contact location");
+      return false;
+    }
+
+    q.set(w,x,y,z);
+
+    inFile >> x >> y >> z;
+    if(inFile.fail()) {
+      DBGA("VirtualContactOnObject::readFromFile - Failed to read virtual contact orientation");
+      return false;
+    }
+
+    t.set(x,y,z);
+    loc = position(x,y,z);
+    frame.set(q,t);
+
+    //normal
+    inFile >> x >> y >> z;
+    if(inFile.fail()) {
+      DBGA("VirtualContactOnObject::readFromFile - Failed to read virtual contact normal");
+      return false;
+    }
+
+    normal.set(x,y,z);
+
+    //sCof
+    inFile >> w;
+    if(inFile.fail()) {
+      DBGA("VirtualContactOnObject::readFromFile - Failed to read virtual contact normal");
+      return false;
+    }
+    sCof = w;
+    return true;
+}
+
+#ifdef ARIZONA_PROJECT_ENABLED
+void
+VirtualContactOnObject::readFromRawData(ArizonaRawExp* are, QString file, int index, bool flipNormal)
+{
+
+    float v;
+    FILE *fp = fopen(file.latin1(), "r");
+    if (!fp) {
+        fprintf(stderr,"Could not open filename %s\n",file.latin1());
+        return;
+    }
+
+    //numFCVectors
+    if(fscanf(fp,"%d",&numFrictionEdges) <= 0) {
+      DBGA("VirtualContactOnObject::readFromRawData - Failed to read virtual contact orientation");
+      return;
+    }
+
+    //frictionEdges
+    for (int i=0; i<numFrictionEdges; i++) {
+        for (int j=0; j<6; j++) {
+          if(fscanf(fp,"%f",&v) <= 0)
+            {
+              DBGA("VirtualContactOnObject::readFromRawData - Failed to read number of friction edges for virtual contacts");
+              return;
+            }
+          frictionEdges[6*i+j] = v;
+        }
+    }
+
+    Quaternion q;
+    vec3 t;
+    q = are->getQuaternion(index);
+    t = are->getContact(index);
+    loc = position(t.x(),t.y(),t.z());
+    frame.set(q,t);
+
+    //normal
+    normal = are->getNormal(index);
+    if(flipNormal){
+        std::cout << "flipped normal" << std::endl;
+        normal = - normal;
+    }
+
+    //sCof, need further consideration
+    sCof = 0.5;
+
+    fclose(fp);
+
+}
+#endif
+
+void
+VirtualContactOnObject::writeToFile(std::ofstream& outFile){
+
+    if (!outFile.is_open())
+    {
+        DBGA("VirtualContactOnObject::writeToFile: failed to open file");
+        return;
+    }
+
+    outFile << numFrictionEdges << std::endl;
+
+    //frictionEdges
+    for (int i=0; i<numFrictionEdges; i++) {
+        for (int j=0; j<6; j++)
+            outFile << frictionEdges[6*i+j] << " ";
+        outFile << std::endl;
+    }
+
+    //frame
+    Quaternion q = frame.rotation();
+    vec3 t = frame.translation();
+    outFile << q.w << " " << q.x << " " << q.y << " " << q.z << std::endl;
+
+    //normal
+    outFile << normal.x() << " " << normal.y() << " " << normal.z() << std::endl;
+
+    //sCof
+    outFile << sCof << std::endl;
+}
