+++ src/contactSetting.cpp
@@ -25,6 +25,9 @@
 
 #include "contactSetting.h"
 #include "contact.h"
+#include "softContact.h"
+#include "virtualContact.h"
+#include "pointContact.h"
 #include "matvec3D.h"
 #include "world.h"
 #include "body.h"
