+++ src/robot.cpp
@@ -50,6 +50,7 @@
 #include "eigenGrasp.h"
 #include "matrix.h"
 #include "tinyxml.h"
+#include "virtualContact.h"
 
 #ifdef USE_DMALLOC
 #include "dmalloc.h"
